+++ src/zepben/eas/client/eas_client.py
@@ -114,11 +114,12 @@ class EasClient:
         """
         Uploads a new study to the Evolve App Server
         :param study: An instance of a data class representing a new study
+        :return: The HTTP response received from the Evolve App Server after attempting to upload the study
         """
         with warnings.catch_warnings():
             if self.__verify_certificate is False:
                 warnings.filterwarnings("ignore", category=InsecureRequestWarning)
-            response = requests.post(
+            return requests.post(
                 construct_url(
                     protocol=self.__protocol,
                     host=self.__host,
@@ -162,7 +163,3 @@ class EasClient:
                 },
                 verify=self.__verify_certificate
             )
-            if not response.ok:
-                warnings.warn(f"Attempted to upload study; "
-                              f"Evolve App Server responded with {response.status_code}: {response.text}")
-            return response